split smart table into multiple files
