+++ mc-react-components/src/lib/smart-table/index.ts
@@ -1,1 +1,2 @@
 export * from './smart-table';
+export * from './smart-table-types';
